Move StatBlockHeader into its own file
